requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
alembic
asyncpg
pydantic
pydantic[email]
passlib[bcrypt]
bcrypt==4.0.1
httpx
requests
PyJWT
structlog
celery[redis]
redis
email-validator
authlib
itsdangerous
google-auth
google-auth-oauthlib
sqladmin[full]
